+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 * Enable Symfony 7
 * Remove support for CommonMark 1
+* Integrate `exists` method into `FrontMatterInterface`
 
 1.5.0 (2022-12-12)
 ------------------